Create primes decoder
The number 72216017 has the word GHOST encoded. The original sequence of prime numbers is: if 2 is the first prime, it corresponds to the letter A, if 3 is the second prime corresponds to the letter B... So on and so on, the primes which 72216017 consists of are 17,19,47,67 and 71 and each number must receive the letters G, H, O, S and T correspondingly. However, I don't know how to make this mapping right. Can somebody help me? I will appreciate it greatly!